all fixes
