+++ CourseApp.Tests/MyAgeTest.cs
@@ -6,21 +6,26 @@ namespace CourseApp.Tests
     public class MyAgeTest
     {
         [Fact]
-        public void NormInt()
-        {
-            Assert.Equal("Вам 18 лет, 10 месяцев и 1 дня", Program.MyAge(2001, 2, 5));
-        }
-
-        [Fact]
         public void NormDate()
         {
-            Assert.Equal("Вам 18 лет, 10 месяцев и 1 дня", Program.MyAge(new DateTime(2001, 2, 5)));
+            Assert.Equal(new DateTime(19, 11, 4), MyAgeClass.DateCompar(new DateTime(2001, 2, 5), new DateTime(2019, 12, 9)));
         }
 
         [Fact]
         public void StrangeDate()
         {
-            Assert.Equal("Вам 0 лет, 0 месяцев и 0 дня", Program.MyAge(DateTime.Now));
+            try
+            {
+                MyAgeClass.DateCompar(DateTime.Now, DateTime.Now);
+            }
+            catch (AgeException ex)
+            {
+                if (ex.Message == "Автору 0 лет")
+                {
+                    Console.WriteLine(ex.Message);
+                    Assert.True(true);
+                }
+            }
         }
 
         [Fact]
@@ -28,7 +33,7 @@ namespace CourseApp.Tests
         {
             try
             {
-                Program.MyAge(2037, 1, 1);
+                MyAgeClass.MyAge(2037, 1, 1);
             }
             catch (AgeException ex)
             {
@@ -42,7 +47,7 @@ namespace CourseApp.Tests
         {
             try
             {
-                Program.MyAge(new DateTime(2037, 1, 1));
+                MyAgeClass.MyAge(new DateTime(2037, 1, 1));
             }
             catch (AgeException ex)
             {
