+++ CourseApp/Program.cs
@@ -64,44 +64,8 @@ namespace CourseApp
                 Console.WriteLine(item.Relax());
             }
 
-            Console.WriteLine(MyAge());
+            Console.WriteLine(MyAgeClass.DateCompar(new DateTime(2001, 2, 5), new DateTime(2019, 12, 9)));
             Console.ReadKey();
-        }
-
-        public static string MyAge()
-        {
-            Console.WriteLine("Введите год своего рождения:");
-            int years = int.Parse(Console.ReadLine());
-            Console.WriteLine("Введите месяц своего рождения:");
-            int months = int.Parse(Console.ReadLine());
-            Console.WriteLine("Введите день своего рождения:");
-            int days = int.Parse(Console.ReadLine());
-            DateTime dayOfBirth = new DateTime(years, months, days);
-            DateTime result = new DateTime(DateTime.Now.Ticks - dayOfBirth.Ticks);
-            return $"Вам {result.Year - 1} лет, {result.Month - 1} месяцев и {result.Day - 1} дня";
-        }
-
-        public static string MyAge(int y, int m, int d)
-        {
-            DateTime dayOfBirth = new DateTime(y, m, d);
-            if (dayOfBirth.Ticks < DateTime.Now.Ticks)
-            {
-                DateTime result = new DateTime(DateTime.Now.Ticks - dayOfBirth.Ticks);
-                return $"Вам {result.Year - 1} лет, {result.Month - 1} месяцев и {result.Day - 1} дня";
-            }
-
-            throw new AgeException("Вы ещё не родились, ожидайте");
-        }
-
-        public static string MyAge(DateTime date)
-        {
-            if (date.Ticks < DateTime.Now.Ticks)
-            {
-                DateTime result = new DateTime(DateTime.Now.Ticks - date.Ticks);
-                return $"Вам {result.Year - 1} лет, {result.Month - 1} месяцев и {result.Day - 1} дня";
-            }
-
-            throw new AgeException("Вы ещё не родились, ожидайте");
         }
     }
 }
